wechall - applied requested changes
