+++ dojo_plugin/api/__init__.py
@@ -7,7 +7,6 @@ from .v1.scoreboard import scoreboard_namespace
 from .v1.ssh_key import ssh_key_namespace
 from .v1.dojo import dojo_namespace
 from .v1.belts import belts_namespace
-from .v1.validate import validate_namespace
 from .v1.score import score_namespace
 
 
@@ -20,5 +19,4 @@ api_v1.add_namespace(scoreboard_namespace, "/scoreboard")
 api_v1.add_namespace(ssh_key_namespace, "/ssh_key")
 api_v1.add_namespace(dojo_namespace, "/dojo")
 api_v1.add_namespace(belts_namespace, "/belts")
-api_v1.add_namespace(validate_namespace, "/validate")
 api_v1.add_namespace(score_namespace, "/score")
