+++ dojo_plugin/api/v1/score.py
@@ -2,36 +2,27 @@ from flask import request
 from flask_restx import Namespace, Resource
 from CTFd.cache import cache
 from CTFd.models import Users
+from CTFd.utils.decorators import ratelimit
 
 from ...models import Dojos
 
 score_namespace = Namespace("score")
 
-def get_user_score(username):
-    # Alternative: hardcode max count and challenge id's for speed
-    # POSSIBLE BUG HERE - topics and courses dojos might result in doubling
-    # but since solves carry over, it should not be an issue
-    # Maybe have a proper scoring system where babyarch has more weight than babysuid
-    user = Users.query.filter_by(name=username).first()
-    if not user:
-        return {"error": "user does not exist"}, 400
+@score_namespace.route("/validate")
+class ValidateUser(Resource):
+    """
+    /validate?username=user&email=user@test.com
+    This endpoint is publicly available, no auth needed.
+    Returns 1 or 0 depending on if username/email exists or not.
+    """
+    @ratelimit(method="GET", limit=10, interval=60)
+    def get(self):
+        username = request.args.get("username")
+        email = request.args.get("email")
+        if not username or not email:
+            return {"error": "`username` and `email` parameters are required"}, 400
 
-    official_dojos = Dojos.query.filter_by(official=1).all()
-    official_challenge_ids = []
-    user_score = 0
-    max_score = 0
-
-    for dojo in official_dojos:
-        for challenge in dojo.challenges:
-            official_challenge_ids.append(challenge.challenge_id)
-
-        max_score += len(dojo.challenges)
-
-    for solve in user.solves:
-        if solve.challenge_id in official_challenge_ids:
-            user_score += 1
-
-    return f"{user_score}:{max_score}"
+        return int(bool(Users.query.filter_by(name=username, email=email).first()))
 
 @score_namespace.route("")
 class ScoreUser(Resource):
@@ -45,4 +36,12 @@ class ScoreUser(Resource):
         if not username:
             return {"error": "`username` parameter is required"}, 400
 
-        return get_user_score(username=username)
+        user = Users.query.filter_by(name=username).first()
+        if not user:
+            return {"error": "user does not exist"}, 400
+
+        official_challenge_ids = {challenge.challenge_id for dojo in Dojos.query.filter_by(official=1).all() for challenge in dojo.challenges}
+        user_score = sum(1 for solve in user.solves if solve.challenge_id in official_challenge_ids)
+        max_score = len(official_challenge_ids)
+
+        return f"{user_score}:{max_score}"
